changed back

+++ src/exp_ms.py
@@ -192,7 +192,7 @@ def append_prev_info(x_it_, scalar_list):
         print("stype: ", s.type())
         print("x_it_ type: ", x_it_.type())
         x_it_ = torch.cat(
-            [x_it_, s.type(torch.DoubleTensor).view(tensor_length(s))]
+            [x_it_, s.type(torch.FloatTensor).view(tensor_length(s))]
         )
     return x_it_
